docs(storybook): move PhoneInput stories to separate files, update storybook version
