+++ package.json
@@ -42,8 +42,8 @@
     "build": "tsup && tsc --emitDeclarationOnly",
     "test": "jest",
     "test:coverage": "jest --coverage",
-    "storybook": "sb dev -p 6006",
-    "storybook:build": "sb build",
+    "storybook": "storybook dev -p 6006",
+    "storybook:build": "storybook build",
     "storybook:preview": "npx http-server ./storybook-static",
     "storybook-test:chrome": "npx playwright test ./.storybook/browser.test --headed --browser=chromium",
     "storybook-test:firefox": "npx playwright test ./.storybook/browser.test --headed --browser=firefox",
@@ -81,10 +81,10 @@
     "@semantic-release/github": "^8.0.6",
     "@semantic-release/npm": "^9.0.1",
     "@semantic-release/release-notes-generator": "^10.0.3",
-    "@storybook/addon-controls": "^7.3.0",
+    "@storybook/addon-controls": "^7.4.1",
     "@storybook/preset-scss": "^1.0.3",
-    "@storybook/react": "^7.3.0",
-    "@storybook/react-vite": "^7.3.0",
+    "@storybook/react": "^7.4.1",
+    "@storybook/react-vite": "^7.4.1",
     "@swc/core": "^1.3.76",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^12.1.5",
@@ -123,7 +123,7 @@
     "react-hook-form": "^7.41.1",
     "sass": "^1.55.0",
     "semantic-release": "^19.0.5",
-    "storybook": "^7.3.0",
+    "storybook": "^7.4.1",
     "stylelint": "^14.13.0",
     "stylelint-config-idiomatic-order": "^9.0.0",
     "stylelint-config-prettier": "^9.0.3",
