+++ src/stories/Dev.stories.tsx
@@ -9,7 +9,7 @@ import { MuiPhone } from './UiLibsExample/components/MuiPhone';
 
 export default {
   title: 'Dev',
-  includeStories: [], // Comment this line to show stories
+  // includeStories: [], // Comment this line to show stories
 } as Meta;
 
 const Title: React.FC<{
