+++ src/stories/PhoneInput/PhoneInput.stories.tsx
@@ -1,0 +1,42 @@
+import type { Meta, StoryObj } from '@storybook/react';
+
+import { PhoneInput } from '../../index';
+
+const meta: Meta<typeof PhoneInput> = {
+  title: 'PhoneInput',
+  component: PhoneInput,
+};
+
+export default meta;
+export type PhoneInputStory = StoryObj<typeof PhoneInput>;
+
+/* eslint-disable simple-import-sort/imports */
+import { Default } from './stories/Default.story';
+import { WithInitialValue } from './stories/WithInitialValue.story';
+import { ForcedDialCode } from './stories/ForcedDialCode.story';
+import { DisabledPrefill } from './stories/DisabledPrefill.story';
+import { DisabledCountryGuess } from './stories/DisabledCountryGuess.story';
+import { WithoutDropdown } from './stories/WithoutDropdown.story';
+import { HiddenDialCode } from './stories/HiddenDialCode.story';
+import { WithCodePreview } from './stories/WithCodePreview.story';
+import { CustomStyles } from './stories/CustomStyles.story';
+import { OnlyBalticCountries } from './stories/OnlyBalticCountries.story';
+import { WithAutofocus } from './stories/WithAutofocus.story';
+import { E164Format } from './stories/E164Format.story';
+import { ControlledMode } from './stories/ControlledMode.story';
+import { CustomFlags } from './stories/CustomFlags.story';
+
+export const _Default = Default;
+export const _WithInitialValue = WithInitialValue;
+export const _ForcedDialCode = ForcedDialCode;
+export const _DisabledPrefill = DisabledPrefill;
+export const _DisabledCountryGuess = DisabledCountryGuess;
+export const _WithoutDropdown = WithoutDropdown;
+export const _HiddenDialCode = HiddenDialCode;
+export const _WithCodePreview = WithCodePreview;
+export const _CustomStyles = CustomStyles;
+export const _OnlyBalticCountries = OnlyBalticCountries;
+export const _WithAutofocus = WithAutofocus;
+export const _E164Format = E164Format;
+export const _ControlledMode = ControlledMode;
+export const _CustomFlags = CustomFlags;
