+++ src/stories/PhoneInput/stories/Default.story.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+
+import { PhoneInput } from '../../../index';
+import { PhoneInputStory } from '../PhoneInput.stories';
+
+export const Default: PhoneInputStory = {
+  render: (args) => <PhoneInput {...args} />,
+};
