+++ src/stories/PhoneInput/stories/DisabledCountryGuess.story.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+import { PhoneInput } from '../../../index';
+import { PhoneInputStory } from '../PhoneInput.stories';
+
+export const DisabledCountryGuess: PhoneInputStory = {
+  name: 'Disabled Country Guess',
+  render: (args) => <PhoneInput {...args} />,
+  args: {
+    defaultCountry: 'gb',
+    disableCountryGuess: true,
+  },
+};
